0.2.1.7
0.2.1.7

+++ R/makePPTx.R
@@ -11,15 +11,16 @@ mycat=function(...,file="report.Rmd"){
 #' @param filename character name of output file
 #' @param rawDataName The name of the rawData
 #' @param rawDataFile The name of the rawData file which the data are to be read from
+#' @param rmdRemove A logical
 #' @importFrom rmarkdown render powerpoint_presentation
 #' @export
 makePPTx=function(data,preprocessing="",filename="report.pptx",rawDataName=NULL,
-                  rawDataFile="rawData.RDS") {
+                  rawDataFile="rawData.RDS",rmdRemove=TRUE) {
 
     file.create("report.Rmd")
     tempReport <-  "report.Rmd"
 
-    title="Conditional Process Anlaysis"
+    title="Conditional Process Analysis"
     subtitle="prepared by processR package"
 
     mycat("---\ntitle: '",title,"'\n")
@@ -33,8 +34,7 @@ makePPTx=function(data,preprocessing="",filename="report.pptx",rawDataName=NULL,
     mycat("```{r setup, include=FALSE}\n")
     mycat("knitr::opts_chunk$set(echo = FALSE,comment=NA,message=FALSE,
           warning=FALSE,
-          fig.width=7,fig.height=5, dpi=300,
-          out.width='80%',fig.align='center')\n")
+          fig.width=7,fig.height=5, dpi=300)\n")
     mycat("```\n")
 
     mycat("```{r}\n")
@@ -65,8 +65,8 @@ makePPTx=function(data,preprocessing="",filename="report.pptx",rawDataName=NULL,
 
     count=nrow(data)
     for(i in 1:count){
-        if(data$title[i]=="") mycat("\n\n----\n\n")
-        else mycat("\n\n#### ",data$title[i],"\n")
+        # if(data$title[i]=="") mycat("\n\n----\n\n")
+        # else mycat("\n\n#### ",data$title[i],"\n")
         mycat("```{r}\n")
         mycat(data$code[i],"\n")
         mycat("```\n\n\n")
@@ -75,56 +75,48 @@ makePPTx=function(data,preprocessing="",filename="report.pptx",rawDataName=NULL,
 
     rmarkdown::render('report.Rmd', output_file="report.pptx",
                       rmarkdown::powerpoint_presentation())
-    file.remove("report.Rmd")
+    if(rmdRemove) file.remove("report.Rmd")
 
 }
+
+
 
 #' Make powerpoint presentation from R file
 #' @param file source file name
 #' @param filename destination file name
+#' @param keyword A string vector
+#' @param rmdRemove A logical
 #' @importFrom readr read_file
-#' @importFrom stringr str_replace str_replace_all
+#' @importFrom stringr str_detect
 #' @export
-r2pptx=function(file,filename="report.pptx"){
+r2pptx=function(file,filename="report.pptx",
+                keyword=c("Diagram","Model","Plot","plot","Table","summary"),
+                rmdRemove=TRUE){
 
-    title<-code<-c()
+    code<-c()
     count=0
 
     text=readr::read_file(file)
     text=unlist(strsplit(text,"\n"))
 
     tempcode=""
+    pattern=paste0(keyword,collapse="|")
+    pattern
+    temp=""
     for( i in seq_along(text)){
-        temp=text[i]
-        temp
-        if(length(grep("---",temp))>0) {
-            if(i>1){
-                code=c(code,tempcode)
-                tempcode=""
-            }
-            title=c(title,"")
-        } else if(length(grep("#",temp))==0) {
-            if(tempcode=="") tempcode=temp
-            else tempcode=paste(tempcode,temp,sep="\n")
+        if(temp=="") {
+            temp=text[i]
         } else{
-            if(grep("#",temp)==1){
-                if(i>1){
-                    code=c(code,tempcode)
-                    tempcode=""
-                }
-                temp=str_replace_all(temp,"#","")
-                temp=str_replace(temp,"^ ","")
-                title=c(title,temp)
-            } else{
-                if(tempcode=="") tempcode=temp
-                else tempcode=paste(tempcode,temp,sep="\n")
-            }
-
+            temp=paste0(temp,text[i],sep="\n")
+        }
+        if(str_detect(temp,pattern)) {
+            code=c(code,temp)
+            temp=""
         }
     }
-    code=c(code,tempcode)
-    df=data.frame(title,code,stringsAsFactors = FALSE)
-    makePPTx(df,filename=filename)
+    code
+    df=data.frame(code,stringsAsFactors = FALSE)
+    makePPTx(df,filename=filename,rmdRemove=rmdRemove)
 }
